+++ doc/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Highlighted Features
 
+* Treat unstemmed words as normal search terms. (Matthew Somerville)
 * Update `/alaveteli_pro` base path to `/pro` (Alexander Griffen, Graeme
   Porteous)
 * Change use of `/alaveteli_pro/info_requests/{request}` to instead appear as